Moved lecture materials into lectures folder
Touched up and cleared up individual.md

A lot of what I touched up in here was trying to
give a clearer explanation of the differences between what
we were doing in parts 1 and 2 versus parts 3 and 4. This
for the most part amounted to trying to provide clearer
explanations at the beginning of parts 1 and parts 3. I also
tried to condense and clean up some things where possible
(see the Spark notes section at the beginning of Part 2, and
the tmux intro at the beginning of part 3).

In a couple of differnet places, I tried to give clearer
explanations of the way Spark works under the hood (see the RDD
paragraph right at the beginning of Part 1 number 1), and what
some general best practices with Spark are (see use of the
collect method in Part 1-number 4, the brief discussion
of locally testing code at the beginning of Part 3, and the mention
of working with a sub-sample in Part 4-number 3).

I renamed the data/toy_data.txt to data/cookie_data.txt,
mainly because its more fun, but also because we say later in
the assignment that these are numbers of cookies bought by people.

A number of things will show up as changes here, when in fact I just
reformatted lines (I was working in vim and 150+ character lines
are no fun). So you might see changes in the git diffs where it
doesn't look like anything was done, and thats probably just a
line that I reformatted to be around ~80 characters.

Clean up and clarify afternoon assignment

The majority of this commit is really just cleaning up
here and there. A lot of the changes aren't real changes -
I was working on vim and long lines are a pain.

The biggest change is probably in the discussion of Word 2 Vec
in the extra-credit. I found a resource that explained it
very clearly (which I linked), and so I added that and re-worked
some of the explanations.

Minor final changes to individual.md

Fix links in pair.md

Fix image links in sparkui.md

Fix image links in lecture_notes.md

Update image link in summary.md

Upload sallamander lectures

Upload data/cookie.txt

Not sure how/where this got lost, but I must have accidentally
kicked it out of the folder when I was reworking things.

Update individual.md instructions

In our solutions we use .groupByKey() in Part 4 - Number 8, but we tell students to use .reduceByKey().
I think it's possible to use either, but for now I'm changing this to match the solutions.

Update sall lectures

Fix typos

The numbering in the assignment was off, but github was rendering our
numbers correctly (i.e. fixing our numbering error, which actually made things
not make sense). Now our numbers are actually correct and everything lines up.

Update sall afternoon lecture